+++ Formula/curry.rb
@@ -6,9 +6,9 @@ class Curry < Formula
   # Update url + sha256 after tagging a release:
   #   git tag v1.2.6 && git push origin v1.2.6
   #   curl -L https://github.com/deconstructo/curry/archive/refs/tags/v1.6.2.tar.gz | shasum -a 256
-  url "https://github.com/deconstructo/curry/archive/refs/tags/v1.9.1.tar.gz"
-  sha256 "b1ae6d780003a9c6aec204f4575fc58fdb64440694e5a3e99516959c67699495"
-  version "1.9.1"
+  url "https://github.com/deconstructo/curry/archive/refs/tags/v1.10.0.tar.gz"
+  sha256 "0000000000000000000000000000000000000000000000000000000000000"
+  version "1.10.0"
 
   head "https://github.com/deconstructo/curry.git", branch: "main"